# services/facade/openapi.yaml
openapi: 3.0.0
info:
  title: VSB EInk Facade Service API
  version: 1.0.0
  description: An API for managing panels, users, and schedules in the VSB EInk Facade Service.
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    tokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HttpError:
      type: object
      properties:
        statusCode:
          type: number
        code:
          type: string
        error:
          type: string
        message:
          type: string
...
    Panel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkedPanelGroup'
      required: [id, name, groups]
    InsertablePanel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkablePanelGroup'
      required: [id]
    UpdatablePanel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/LinkablePanelGroup'
    LinkablePanel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id]
    LinkedPanel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          readOnly: true
      required: [ id, name ]
    PanelGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        panels:
          type: array
          items:
            $ref: '#/components/schemas/LinkedPanel'
        managedBy:
          type: array
          items:
            $ref: '#/components/schemas/LinkedUserGroup'
      required: [id, name, panels, managedBy]
    InsertablePanelGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        panels:
          type: array
          items:
            $ref: '#/components/schemas/LinkablePanel'
        managedBy:
          type: array
          items:
            $ref: '#/components/schemas/LinkableUserGroup'
      required: [id]
    UpdatablePanelGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        panels:
          type: array
          items:
            $ref: '#/components/schemas/LinkablePanel'
        managedBy:
          type: array
          items:
            $ref: '#/components/schemas/LinkableUserGroup'
    LinkablePanelGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id]
    LinkedPanelGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          readOnly: true
      required: [ id, name ]
...
    LinkableUserGroup:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
      required: [id]
    LinkedUserGroup:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          readOnly: true
      required: [id, name]
...
    ScheduledJob:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
          nullable: true
        cron:
          type: string
        target:
          type: string
        command:
          type: string
        content:
          type: array
          items:
            type: string
        precise:
          type: boolean
        priority:
          type: number
        cycle:
          type: number
        shouldCycle:
          type: boolean
        disabled:
          type: boolean
      required: [ id, name, description, cron, target, command, content, precise, priority, cycle, shouldCycle, disabled]
    InsertableScheduledJob:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        cron:
          type: string
        target:
          type: string
        command:
          type: string
        content:
          type: array
          items:
            type: string
        priority:
          type: number
        shouldCycle:
          type: boolean
        disabled:
          type: boolean
      required: [name, cron, target, command]
    UpdatableScheduledJob:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        cron:
          type: string
        target:
          type: string
        command:
          type: string
        content:
          type: array
          items:
            type: string
        priority:
          type: number
        cycle:
          type: number
        shouldCycle:
          type: boolean
        disabled:
          type: boolean
  parameters:
...
    panelId:
      name: panelId
      in: path
      required: true
      schema:
        type: string
    panelGroupId:
      name: panelGroupId
      in: path
      required: true
      schema:
        type: string
...
    jobId:
      name: jobId
      in: path
      required: true
      schema:
        type: number
...
    jobTarget:
      name: target
      in: query
      required: false
      schema:
        type: string
    jobHasSeconds:
      name: hasSeconds
      in: query
      required: false
      schema:
        type: boolean
    jobDisabled:
      name: disabled
      in: query
      required: false
      schema:
        type: boolean
        default: false
...
paths:
...
  /panels/:
    get:
      operationId: getPanels
      tags: [panels]
      security:
        - tokenAuth: [PANELS_READ]
      responses:
        '200':
          description: List of panels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Panel'
    post:
      operationId: createPanel
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertablePanel'
      responses:
        '201':
          description: Panel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Panel'
  /panels/{panelId}:
    get:
      operationId: getPanel
      tags: [panels]
      security:
        - tokenAuth: [PANELS_READ]
      parameters:
        - $ref: '#/components/parameters/panelId'
      responses:
        '200':
          description: Requested panel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Panel'
        '404':
          description: Panel with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    patch:
      operationId: updatePanel
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatablePanel'
      parameters:
        - $ref: '#/components/parameters/panelId'
      responses:
        '200':
          description: Panel updated
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Panel'
    delete:
      operationId: deletePanel
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      parameters:
        - $ref: '#/components/parameters/panelId'
      responses:
        '200':
          description: Panel deleted
  /panel-groups/:
    get:
      operationId: getPanelGroups
      tags: [panels]
      security:
        - tokenAuth: [PANELS_READ]
      responses:
        '200':
          description: List of panel groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PanelGroup'
    post:
      operationId: createPanelGroup
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertablePanelGroup'
      responses:
        '201':
          description: Panel group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PanelGroup'
  /panel-groups/{panelGroupId}:
    get:
      operationId: getPanelGroup
      tags: [panels]
      security:
        - tokenAuth: [PANELS_READ]
      parameters:
        - $ref: '#/components/parameters/panelGroupId'
      responses:
        '200':
          description: Requested panel group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PanelGroup'
        '404':
          description: Panel group with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    patch:
      operationId: updatePanelGroup
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatablePanelGroup'
      parameters:
        - $ref: '#/components/parameters/panelGroupId'
      responses:
        '200':
          description: Panel group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PanelGroup'
        '404':
          description: Panel group with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
    delete:
      operationId: deletePanelGroup
      tags: [panels]
      security:
        - tokenAuth: [PANELS_WRITE]
      parameters:
        - $ref: '#/components/parameters/panelGroupId'
      responses:
        '204':
          description: Panel group deleted
        '404':
          description: Panel group with the given ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
...
  /schedule/:
    get:
      operationId: getScheduledJobs
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_READ]
      parameters:
        - $ref: '#/components/parameters/jobTarget'
        - $ref: '#/components/parameters/jobHasSeconds'
        - $ref: '#/components/parameters/jobDisabled'
      responses:
        '200':
          description: List of scheduled jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduledJob'
    patch:
      operationId: updateScheduledJobs
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UpdatableScheduledJob'
      responses:
        '204':
          description: Scheduled jobs updated
    post:
      operationId: createScheduledJob
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_WRITE]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertableScheduledJob'
      responses:
        '201':
          description: Scheduled job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
  /schedule/{jobId}:
    get:
      operationId: getScheduledJob
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_READ]
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '200':
          description: Requested scheduled job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
    patch:
      operationId: updateScheduledJob
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_WRITE]
      parameters:
        - $ref: '#/components/parameters/jobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatableScheduledJob'
      responses:
        '200':
          description: Scheduled job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledJob'
    delete:
      operationId: deleteScheduledJob
      tags: [schedule]
      security:
        - tokenAuth: [SCHEDULE_WRITE]
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        '204':
          description: Scheduled job deleted
...